+++ src/senaite/fhir/resource/device.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+
+from senaite.fhir.interfaces import IDeviceResource
+from senaite.fhir.resource import FHIRResource
+from zope.interface import implementer
+
+
+@implementer(IDeviceResource)
+class DeviceResource(FHIRResource):
+    pass